% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused.ridge-package.R
\docType{package}
\name{fused.ridge-package}
\alias{fused.ridge}
\alias{fused.ridge-package}
\title{fused.ridge: What the Package Does (One Line, Title Case)}
\description{
Estimate ridge regression with fused penalty.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pgg1309/fused.ridge}
  \item Report bugs at \url{https://github.com/pgg1309/fused.ridge/issues}
}

}
\author{
\strong{Maintainer}: Paulo Gustavo Grahl \email{pgrahl@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
