% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_model.R
\name{fused_model}
\alias{fused_model}
\title{Create a \code{fused_model}}
\usage{
fused_model(mode = "regression", penalty = NULL)
}
\arguments{
\item{mode}{Needs to be regression}

\item{penalty}{sets the penalty parameter (lambda)}
}
\value{
A \code{fused_model} object.
}
\description{
Create a \code{fused_model}
}
