% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_ridge-fit.R
\name{fused_ridge}
\alias{fused_ridge}
\alias{fused_ridge.default}
\alias{fused_ridge.data.frame}
\alias{fused_ridge.matrix}
\alias{fused_ridge.formula}
\alias{fused_ridge.recipe}
\title{Fit a \code{fused_ridge}}
\usage{
fused_ridge(x, lambda, ...)

\method{fused_ridge}{default}(x, lambda, ...)

\method{fused_ridge}{data.frame}(x, y, lambda, ...)

\method{fused_ridge}{matrix}(x, y, lambda, ...)

\method{fused_ridge}{formula}(formula, data, lambda, ...)

\method{fused_ridge}{recipe}(x, data, lambda, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{lambda}{A numeric value specifying the penalty parameter.}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{fused_ridge} object.
}
\description{
\code{fused_ridge()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]
lambda <- 1

# XY interface
# mod <- fused_ridge(predictors, outcome, lambda)

# Formula interface
# mod2 <- fused_ridge(mpg ~ ., mtcars, lambda)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
# mod3 <- fused_ridge(rec, mtcars, lambda)

}
