% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_ridge-predict.R
\name{predict}
\alias{predict}
\alias{predict.default}
\alias{predict.fused_ridge}
\title{Predict from a \code{fused_ridge}}
\usage{
predict(object, new_data, ...)

\method{predict}{default}(object, new_data, ...)

\method{predict}{fused_ridge}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{fused_ridge} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{...}{Not used, but required for extensibility.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{fused_ridge}
}
\examples{
train <- mtcars[1:20,]
test <- mtcars[21:32, -1]
lambda <- 1

# Fit
#mod <- fused_ridge(mpg ~ cyl + log(drat), train, lambda)

# Predict, with preprocessing
#predict(mod, test)


}
